requirements.txt
pyOpenSSL==22.1.0
requests==2.26.0
